backend/requirements.txt
django
django-cors-headers
djangorestframework
pandas
numpy
python-dotenv